Service to generate and calculate the user's result on the executed quiz

+++ src/main/java/com/fincons/dto/QuizResultsDto.java
@@ -19,7 +19,7 @@ public class QuizResultsDto {
     private UserDto user;
 
     private QuizDto quiz;
-    private double totalScore;
+    private float totalScore;
 
     // TODO List<Integer> lista di indici delle risposte date
